:recycle: Separate ViewGraphImpl
